Reject messages with incorrect length of data (#10)
Messages received from the desk controller are 7 (or more) bytes long.
Nevertheless, on some occasions `deskctl` receives a payload with
incorrect data length (the length exceeds the length of the payload) but
having a correct preamble (`0xf2`) and EoM (`0x7e`).
Consequently, deskctl panics with "index out of range" runtime error.

To fix this regression (until we decipher the payload in question),
`isValidData` function fails when the length of data exceeds the length
of the payload itself.

Closes #9.